+++ packages/@leight/file-client/src/sdk/ClientStore.ts
@@ -43,4 +43,4 @@ export const FileSortStore                    = createSortContext<IFileSortSchem
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_xjpkuecvz45gniwaxh7skpn3 = true;
+export const $leight_ffrfyok0cowy7fb1lcaw61da = true;