+++ packages/@leight/job-client/src/sdk/ClientSourceProvider.tsx
@@ -58,4 +58,4 @@ export const JobQueryProvider: FC<IJobQueryProviderProps> = props => {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_eiko56rs7isny0k7reci9dx4             = true;
+export const $leight_cvdheoy86tgq2yu7iapmd7e1             = true;