+++ packages/@leight/job-server/src/sdk/ServerPrismaSource.ts
@@ -89,4 +89,4 @@ export class JobBasePrismaSource extends AbstractSource<IJobSourceSchema> {
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_qwrjnvkpepzldc9avxgc9ll0 = true;
+export const $leight_dfyxmppxzorrag192uih043c = true;