+++ packages/@leight/job-server/src/sdk/ServerTrpc.ts
@@ -18,4 +18,4 @@ export const JobSourceProcedure               = withSourceProcedure<IJobSourceSc
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_kvoq15f7lr9kirg03mytub2b = true;
+export const $leight_qnihh6lmpdk9ofy2ifypy3ai = true;