+++ packages/@leight/job/src/sdk/Source.ts
@@ -19,4 +19,4 @@ export const JobSourceContext                 = (container: IContainer) => new S
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_s0vygtdt2zoqz22bvvqqdj4q = true;
+export const $leight_dob99diwpcxy7x1iqamtamqp = true;