+++ packages/@leight/user/src/sdk/Source.ts
@@ -20,4 +20,4 @@ export const UserSourceContext                = (container: IContainer) => new S
  * Default export marking a file it's generated and also preventing failing
  * an empty file export (every module "must" have an export).
  */
-export const $leight_woh9s6199h90aff25jice2yk = true;
+export const $leight_s12szjhf3m344jbl6gob70xr = true;