+++ packages/@leight/cursor-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/cursor-client
 
+## 0.5.31
+
+### Patch Changes
+
+- Version bump
+
 ## 0.5.30
 
 ### Patch Changes