+++ packages/@leight/file-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @leight/file-server
 
+## 0.5.70
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.78
+    - @leight/file@0.5.69
+
 ## 0.5.69
 
 ### Patch Changes