+++ packages/@leight/file/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/file
 
+## 0.5.69
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/prisma@0.5.78
+
 ## 0.5.68
 
 ### Patch Changes