+++ packages/@leight/source-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @leight/source-client
 
+## 0.5.32
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+    - @leight/cursor-client@0.5.31
+
 ## 0.5.31
 
 ### Patch Changes